# contracts/openapi/torgnexa-v1.yaml
openapi: 3.1.0
info:
  title: TORGNEXA API
  version: 0.16.0
servers:
  - url: /api/v1
paths:
...
  /lineage/timeline:
    get:
      operationId: getLineageTimeline
      parameters:
        - {name: system, in: query, required: true, schema: {type: string}}
        - {name: entity_type, in: query, required: true, schema: {type: string}}
        - {name: entity_id, in: query, required: true, schema: {type: string}}
        - {name: field, in: query, required: false, schema: {type: string}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
        - {name: before_at, in: query, required: false, schema: {type: string, format: date-time, pattern: 'Z$'}}
        - {name: before_id, in: query, required: false, schema: {type: string}}
      responses:
        '200':
          description: Immutable provenance timeline for one canonical entity/field
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LineageTimelinePage'}
        '400': {description: Invalid lineage query}
        '401': {description: Authenticated tenant scope missing}
...
  /counterparties/search:
    get:
      operationId: searchLegalParties
      parameters:
        - {name: q, in: query, required: false, schema: {type: string, maxLength: 200}}
        - {name: inn, in: query, required: false, schema: {type: string, pattern: '^[0-9]{10,12}$'}}
        - {name: registration_id, in: query, required: false, schema: {type: string, maxLength: 32}}
        - {name: party_type, in: query, required: false, schema: {type: string, enum: [legal_entity, individual_entrepreneur, branch]}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
      responses:
        '200':
          description: Tenant-scoped canonical legal-party search results
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LegalPartySearchPage'}
        '400': {description: Invalid legal-party search query}
        '401': {description: Authenticated tenant scope missing}
...
  /products:
    post:
      operationId: createProduct
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProductWrite'}}}}
      responses: {'201': {description: Product created}, '400': {description: Invalid product}, '409': {description: Duplicate code}}
    get:
      operationId: listProducts
      parameters:
        - {name: q, in: query, required: false, schema: {type: string, maxLength: 256}}
        - {name: status, in: query, required: false, schema: {type: string, enum: [draft, active, archived]}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
        - {$ref: '#/components/parameters/Cursor'}
      responses:
        '200':
          description: Tenant-scoped canonical product search/list page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExternalSearchProductSearchPageV1'}
        '400': {description: Invalid product search query or cursor}
        '401': {description: Authenticated tenant scope missing}
...
  /orders:
    get:
      operationId: listOrders
      parameters:
        - {name: q, in: query, required: false, schema: {type: string, maxLength: 256}}
        - {name: status, in: query, required: false, schema: {type: string, enum: [pending, confirmed, processing, fulfilled, cancelled]}}
        - {name: placed_from, in: query, required: false, schema: {type: string, format: date-time, pattern: 'Z$'}}
        - {name: placed_to, in: query, required: false, schema: {type: string, format: date-time, pattern: 'Z$'}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
        - {$ref: '#/components/parameters/Cursor'}
      responses:
        '200':
          description: Tenant-scoped canonical order search/list page
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExternalSearchOrderSearchPageV1'}
        '400': {description: Invalid order search query or cursor}
        '401': {description: Authenticated tenant scope missing}
  /orders/demo:
    post:
      operationId: createDemoOrders
      description: Idempotently creates a synthetic tenant-scoped dataset with five orders, product, offer, warehouse, stock position, verified compliance declaration and personal demo notifications for the authenticated actor, then appends audit evidence.

      parameters:
        - {name: Idempotency-Key, in: header, required: false, schema: {type: string, minLength: 1, maxLength: 256}}
      responses:
        '201':
          description: Demo dataset created, or already present
          content:
            application/json:
              schema:
                type: object
                required: [created]
                properties:
                  created: {type: integer, minimum: 0, maximum: 5}
        '401': {description: Authenticated tenant scope missing}
        '403': {description: User cannot create demo data}
    delete:
      operationId: deleteDemoOrders
      description: Idempotently hides the synthetic demo dataset while retaining immutable canonical order history and appending audit evidence.
      parameters:
        - {name: Idempotency-Key, in: header, required: false, schema: {type: string, minLength: 1, maxLength: 256}}
      responses:
        '200':
          description: Demo dataset deleted, or was already absent
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [deleted]
                properties:
                  deleted: {type: integer, minimum: 0, maximum: 5}
        '401': {description: Authenticated tenant scope missing}
        '403': {description: Administrator permission required}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: {type: string, maxLength: 2048, pattern: '^v1\.[A-Za-z0-9_-]+$'}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9._:/-]{0,127}$'}
  schemas:
...
    ProductWrite:
      type: object
      properties:
        code: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._:/-]{0,127}$'}
        title: {type: string, minLength: 1, maxLength: 500}
        description: {type: string, maxLength: 20000}
        version: {type: integer, minimum: 1}
        status: {type: string, enum: [draft, active, archived]}
        category_id: {type: string}
...
    LineageTimelinePage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          maxItems: 200
          items: {$ref: '#/components/schemas/LineageRecord'}
        next_cursor:
          type: object
          required: [occurred_at, id]
          properties:
            occurred_at: {type: string, format: date-time, pattern: 'Z$'}
            id: {type: string}
    LineageRecord:
      type: object
      required: [id, source, operation, output, inputs, transformation, correlation_id, audit_id, event_id, result, occurred_at]
      properties:
        id: {type: string}
        source: {type: string}
        actor_id: {type: string}
        operation: {type: string}
        output: {type: object}
        inputs: {type: array, items: {type: object}}
        transformation: {type: object}
        correlation_id: {type: string}
        causation_id: {type: string}
        audit_id: {type: string}
        event_id: {type: string}
        result: {type: string, enum: [applied, observed, rejected]}
        occurred_at: {type: string, format: date-time, pattern: 'Z$'}
    LegalPartySearchPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            type: object
            required: [party_type, party_id, code, display_name, status]
            properties:
              party_type: {type: string, enum: [legal_entity, individual_entrepreneur, branch]}
              party_id: {type: string}
              code: {type: string}
              display_name: {type: string}
              inn: {type: string}
              registration_id: {type: string}
              status: {type: string, enum: [draft, active, archived]}
...
    ExternalSearchOrderSearchPageV1:
      title: Order search page v1
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/ExternalSearchOrderSearchPageV1/$defs/hit'
        next_cursor:
          type: string
          maxLength: 2048
          pattern: ^v1\.[A-Za-z0-9_-]+$
      $defs:
        id:
          type: string
          pattern: ^(?:[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|[0-7][0-9A-HJKMNP-TV-Z]{25})$
        hit:
          type: object
          additionalProperties: false
          required:
          - id
          - order_number
          - status
          - currency
          - grand_minor_units
          - placed_at
          - updated_at
          properties:
            id:
              $ref: '#/components/schemas/ExternalSearchOrderSearchPageV1/$defs/id'
            order_number:
              type: string
              pattern: ^[A-Za-z0-9][A-Za-z0-9._:/-]{0,127}$
            status:
              type: string
              enum:
              - pending
              - confirmed
              - processing
              - fulfilled
              - cancelled
            currency:
              type: string
              pattern: ^[A-Z]{3}$
            grand_minor_units:
              type: integer
              minimum: 0
            placed_at:
              type: string
              format: date-time
              pattern: Z$
            updated_at:
              type: string
              format: date-time
              pattern: Z$
    ExternalSearchProductSearchPageV1:
      title: Product search page v1
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/ExternalSearchProductSearchPageV1/$defs/hit'
        next_cursor:
          type: string
          maxLength: 2048
          pattern: ^v1\.[A-Za-z0-9_-]+$
      $defs:
        id:
          type: string
          pattern: ^(?:[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|[0-7][0-9A-HJKMNP-TV-Z]{25})$
        hit:
          type: object
          additionalProperties: false
          required:
          - id
          - code
          - title
          - status
          - updated_at
          properties:
            id:
              $ref: '#/components/schemas/ExternalSearchProductSearchPageV1/$defs/id'
            code:
              type: string
              pattern: ^[A-Za-z0-9][A-Za-z0-9._:/-]{0,127}$
            title:
              type: string
              minLength: 1
              maxLength: 300
            status:
              type: string
              enum:
              - draft
              - active
              - archived
            updated_at:
              type: string
              format: date-time
              pattern: Z$
...